Moves synchronous context-free grammar and parser code to scfg package.
